KAFKA-2979: Renamed acls.py to kafka_acls.py.
